# optimize rope function to avoid call powf in the loop