This slightly improves the perf.

Before:
llama_print_timings:        load time =  6973.64 ms
llama_print_timings:      sample time =   345.88 ms /   128 runs   (    2.70 ms per run)
llama_print_timings: prompt eval time =  6227.73 ms /    29 tokens (  214.75 ms per token)
llama_print_timings:        eval time = 34661.36 ms /   127 runs   (  272.92 ms per run)
llama_print_timings:       total time = 41992.88 ms

After:
llama_print_timings:        load time =  7668.98 ms
llama_print_timings:      sample time =   345.03 ms /   128 runs   (    2.70 ms per run)
llama_print_timings: prompt eval time =  6906.97 ms /    29 tokens (  238.17 ms per token)
llama_print_timings:        eval time = 33445.33 ms /   127 runs   (  263.35 ms per run)
llama_print_timings:       total time = 41471.40 ms