@j-f1 2023-04-06
Can you re-run the tests a few times (like a total 3–5 for each) and confirm that the improvement is not just a result of random noise in the amount of time the process takes to complete?