@howard0su 2023-04-06
yes, I actually did that already. the above is one sample:

Before:
```
llama_print_timings:        load time =  6508.98 ms
llama_print_timings:      sample time =    48.19 ms /    18 runs   (    2.68 ms per run)
llama_print_timings: prompt eval time =  5758.73 ms /    26 tokens (  221.49 ms per token)
llama_print_timings:        eval time =  4487.70 ms /    17 runs   (  263.98 ms per run)
llama_print_timings:       total time = 11046.59 ms

llama_print_timings:        load time =  6430.18 ms
llama_print_timings:      sample time =   306.51 ms /   115 runs   (    2.67 ms per run)
llama_print_timings: prompt eval time =  5708.87 ms /    26 tokens (  219.57 ms per token)
llama_print_timings:        eval time = 30756.65 ms /   114 runs   (  269.80 ms per run)
llama_print_timings:       total time = 37504.31 ms

llama_print_timings:        load time =  6422.55 ms
llama_print_timings:      sample time =   346.25 ms /   128 runs   (    2.71 ms per run)
llama_print_timings: prompt eval time =  5664.71 ms /    26 tokens (  217.87 ms per token)
llama_print_timings:        eval time = 33817.97 ms /   127 runs   (  266.28 ms per run)
llama_print_timings:       total time = 40599.00 ms
```

After:
```
llama_print_timings:        load time =  6399.90 ms
llama_print_timings:      sample time =   342.14 ms /   128 runs   (    2.67 ms per run)
llama_print_timings: prompt eval time =  5686.61 ms /    26 tokens (  218.72 ms per token)
llama_print_timings:        eval time = 33954.48 ms /   127 runs   (  267.36 ms per run)
llama_print_timings:       total time = 40708.52 ms

llama_print_timings:        load time =  6352.47 ms
llama_print_timings:      sample time =   342.94 ms /   128 runs   (    2.68 ms per run)
llama_print_timings: prompt eval time =  5649.50 ms /    26 tokens (  217.29 ms per token)
llama_print_timings:        eval time = 33675.62 ms /   127 runs   (  265.16 ms per run)
llama_print_timings:       total time = 40383.40 ms

llama_print_timings:        load time =  6454.62 ms
llama_print_timings:      sample time =   106.15 ms /    40 runs   (    2.65 ms per run)
llama_print_timings: prompt eval time =  5668.35 ms /    26 tokens (  218.01 ms per token)
llama_print_timings:        eval time = 10139.37 ms /    39 runs   (  259.98 ms per run)
llama_print_timings:       total time = 16703.99 ms

```