@howard0su 2023-04-06
If you can share your python script, it will be great. It is a good tool to help some performance experimenting.