@KASR 2023-04-06
> If you can share your python script, it will be great. It is a good tool to help some performance experimenting.

Sure, as a warning it's just a simple loop 😅, you can find the code here --> [benchmark_pr_llama.py](https://gist.github.com/KASR/e2fcc5433b05873e85cc0c0bbe194fc5)