@howard0su 2023-04-13
yes this is not approximate. this is just a math equation.